edit & delete

+++ src/main/java/com/sparta/scheduler/controller/SchedulerController.java
@@ -2,7 +2,10 @@ package com.sparta.scheduler.controller;
 
 import com.sparta.scheduler.dto.SchedulerRequestDto;
 import com.sparta.scheduler.dto.SchedulerResponseDto;
+import com.sparta.scheduler.entity.Schedule;
+import com.sparta.scheduler.repository.SchedulerRepository;
 import com.sparta.scheduler.service.SchedulerService;
+import org.springframework.http.ResponseEntity;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.web.bind.annotation.*;
 
@@ -31,14 +34,31 @@ public class SchedulerController {
     }
 
     @PutMapping("/schedules/{id}")
-    public void updateSchedule(@PathVariable Long id, @RequestBody SchedulerRequestDto requestDto){
-        SchedulerService schedulerService = new SchedulerService(jdbcTemplate);
-        schedulerService.updateSchedule(id, requestDto);
+    public ResponseEntity<?> updateSchedule(@PathVariable Long id, @RequestBody SchedulerRequestDto requestDto) {
+        String inputPassword = requestDto.getPassword();
+        SchedulerRepository schedulerRepository = new SchedulerRepository(jdbcTemplate);
+        Schedule schedule = schedulerRepository.findById(id);
+
+        if (schedule != null && schedule.getPassword().equals(inputPassword)) {
+            SchedulerService schedulerService = new SchedulerService(jdbcTemplate);
+            schedulerService.updateSchedule(id, requestDto);
+            return ResponseEntity.ok("일정 변경에 성공하였습니다.");
+        } else {
+            return ResponseEntity.badRequest().body("비밀번호가 일치하지 않습니다.");
+        }
     }
 
-    @DeleteMapping("/schedules/{id}")
-    public void deleteSchedule(@PathVariable Long id){
+    @RequestMapping(value = "/schedules/{id}", method = RequestMethod.DELETE)
+    public ResponseEntity<?> deleteSchedule(@PathVariable Long id, @RequestParam String password) {
+        SchedulerRepository schedulerRepository = new SchedulerRepository(jdbcTemplate);
         SchedulerService schedulerService = new SchedulerService(jdbcTemplate);
-        schedulerService.deleteSchedule(id);
+        Schedule schedule = schedulerRepository.findById(id);
+        if (schedule != null && schedule.getPassword().equals(password)) {
+            schedulerService.deleteSchedule(id);
+            return ResponseEntity.ok(true);
+        } else {
+            return ResponseEntity.ok(false);
+        }
     }
+
 }
